requirements.txt
certifi==2018.10.15
chardet==3.0.4
Deprecated==1.2.4
idna==2.7
PyGithub==1.43.3
PyJWT==1.6.4
python-dotenv==0.9.1
requests==2.20.1
six==1.11.0
slackbot==0.5.3
slacker==0.9.65
urllib3==1.24.1
websocket-client==0.44.0
wrapt==1.10.11
